fix: missing utils reference
