+++ cortex-js/src/main.ts
@@ -4,7 +4,6 @@ import {
 } from '@/infrastructure/constants/cortex';
 import { getApp } from './app';
 import chalk from 'chalk';
-import { cleanLogs } from './utils/log';
 
 async function bootstrap() {
   const app = await getApp();
@@ -13,8 +12,6 @@ async function bootstrap() {
   const port = process.env.CORTEX_JS_PORT || defaultCortexJsPort;
 
   try {
-    // Clean logs periodically
-    cleanLogs();
     await app.listen(port, host);
     console.log(chalk.blue(`Started server at http://${host}:${port}`));
     console.log(
